a few finishing touches and javadoc comments

+++ src/main/java/wright/firstproject/Controllers/MainController.java
@@ -24,6 +24,9 @@ import java.util.ResourceBundle;
 import static wright.firstproject.Controllers.Inventory.getAllParts;
 import static wright.firstproject.Controllers.Inventory.getAllProducts;
 
+/** This is the class for the main window or first window shown.
+ LOGICAL ERROR To overcome the issue when selecting a part to delete after having searched the object would still display, I made the search term and table reset after deleting an object.
+ FUTURE ENHANCEMENT I already added some basic styling to the buttons of this application, I would like to make this project more aesthetically pleasing in the future.*/
 public class MainController implements Initializable {
     @FXML
     private TableView partsTable, productsTable;
@@ -39,7 +42,7 @@ public class MainController implements Initializable {
     @FXML
     private TextField partSearch, productSearch;
 
-
+    /** This method sets the tables and columns to our Part and Product models.*/
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
         partsTable.setItems(parts);
@@ -64,7 +67,7 @@ public class MainController implements Initializable {
             productSearch(newValue);
         });
     }
-
+    /** Changes the screen to the add part screen */
     public void toAddPart(ActionEvent actionEvent) throws IOException{
         Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/wright/firstproject/add-part.fxml")));
         Stage stage = (Stage)((javafx.scene.Node)actionEvent.getSource()).getScene().getWindow();
@@ -73,7 +76,7 @@ public class MainController implements Initializable {
         stage.setScene(scene);
         scene.getStylesheets().add(MainApplication.class.getResource("bootstrap3.css").toExternalForm());
     }
-
+    /** Changes the screen to the modify part screen */
     public void goToModifyPart(ActionEvent actionEvent) throws IOException{
         try {
             selectedPart = (Part) partsTable.getSelectionModel().getSelectedItem();
@@ -88,15 +91,15 @@ public class MainController implements Initializable {
             System.out.println(error);
         }
     }
-
+    /** gets the part selected in the tableView */
     public static Part getSelectedPart(){
         return selectedPart;
     }
-
+    /** Exits the application on button click */
     public void getOuttaHere() {
         System.exit(0);
     }
-
+    /** changes the screen to the add product screen */
     public void toAddProduct(ActionEvent actionEvent) throws IOException{
         Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/wright/firstproject/add-product.fxml")));
         Stage stage = (Stage)((javafx.scene.Node)actionEvent.getSource()).getScene().getWindow();
@@ -105,7 +108,7 @@ public class MainController implements Initializable {
         stage.setScene(scene);
         scene.getStylesheets().add(MainApplication.class.getResource("bootstrap3.css").toExternalForm());
     }
-
+    /** Changes the screen to the modify product screen */
     public void toModifyProduct(ActionEvent actionEvent) throws IOException{
         try {
             selectedProduct = (Product) productsTable.getSelectionModel().getSelectedItem();
@@ -119,27 +122,34 @@ public class MainController implements Initializable {
             Inventory.warnDialog("No Product Selected", "Please select a product to modify");
         }
     }
-
+    /** method to delete a part */
     public void onDeletePart() {
         selectedPart = (Part) partsTable.getSelectionModel().getSelectedItem();
-        Inventory.deletePart(selectedPart);
-        partsTable.setItems(parts);
-        partSearch.setText("");
+        if (Inventory.confirmDelete("Confirm?", "Are you sure you want to delete " + selectedPart.getName() + " ?")) {
+            Inventory.deletePart(selectedPart);
+            partsTable.setItems(parts);
+            partSearch.setText("");
+        }
     }
-
+    /** method to delete a product */
     public void onDeleteProd() {
         selectedProduct = (Product) productsTable.getSelectionModel().getSelectedItem();
-        Inventory.deleteProduct(selectedProduct);
-        productsTable.setItems(products);
-        productSearch.setText("");
+        if (Inventory.confirmDelete("Confirm?", "Are you sure you want to delete " + selectedProduct.getName() +" ?")){
+            Inventory.deleteProduct(selectedProduct);
+            productsTable.setItems(products);
+            productSearch.setText("");
+        }
     }
-
+    /** method to search for a product.
+     @param searchTerm a string sent to be searched for */
     public void productSearch(String searchTerm) {
         String loweredSearch = searchTerm.toLowerCase();
         productsTable.setItems(Inventory.lookupProduct(loweredSearch));
     }
 //    This function takes the listener from the text field and sets the partsTable data
 //    to the output of the lookupPart Function
+    /** method to search for a part
+     @param searchTerm a string to be searched for */
     public void partSearch(String searchTerm) {
         String loweredSearch = searchTerm.toLowerCase();
         partsTable.setItems(Inventory.lookupPart(loweredSearch));
